CardGameEx - AceLow complete
The Card class is written. The AceLow class is written. This implements
the Comparator interface and is non-generic. It will compare cards and
return a value based on which should be lower. In this case, Ace is low.
The Test class will create a list of cards, then get them sorted AceLow.
Thanks to Michael for sorting out the final bits of syntax.